interesting bug
a kind of interesting bug has occurred as I added some debugging lines to isolate what's gone wrong with the the game and score pointers - where bullets dont delete themselves and enemies (as the debugging lines just return the loop before it can get to that point)- so they just stop when in contact with enemies. Just an interesting effect a bit like they're passing through jelly or something, Could be used for something one day.